+++ scripts/DBUpdate/UpgradeDatabaseStructure/CommunicationLogIndex.pm
@@ -1,0 +1,57 @@
+# --
+# Copyright (C) 2001-2018 OTRS AG, http://otrs.com/
+# --
+# This software comes with ABSOLUTELY NO WARRANTY. For details, see
+# the enclosed file COPYING for license information (AGPL). If you
+# did not receive this file, see http://www.gnu.org/licenses/agpl.txt.
+# --
+
+package scripts::DBUpdate::UpgradeDatabaseStructure::CommunicationLogIndex;    ## no critic
+
+use strict;
+use warnings;
+
+use parent qw(scripts::DBUpdate::Base);
+
+our @ObjectDependencies = ();
+
+=head1 NAME
+
+scripts::DBUpdate::UpgradeDatabaseStructure::CommunicationLogIndex - Adds an index to the communication log table.
+
+=cut
+
+sub Run {
+    my ( $Self, %Param ) = @_;
+
+    my $Verbose = $Param{CommandlineOptions}->{Verbose} || 0;
+
+    my @XMLStrings = (
+
+        # Create an Index for the column 'start_time' in table 'communication_log',
+        #   used in the communication-log purge.
+        '<TableAlter Name="communication_log">
+            <IndexCreate Name="communication_start_time">
+                <IndexColumn Name="start_time"/>
+            </IndexCreate>
+        </TableAlter>',
+    );
+
+    return if !$Self->ExecuteXMLDBArray(
+        XMLArray => \@XMLStrings,
+    );
+
+    return 1;
+}
+
+1;
+
+=head1 TERMS AND CONDITIONS
+
+This software is part of the OTRS project (L<http://otrs.org/>).
+
+This software comes with ABSOLUTELY NO WARRANTY. For details, see
+the enclosed file COPYING for license information (AGPL). If you
+did not receive this file, see L<http://www.gnu.org/licenses/agpl.txt>.
+
+=cut